moved speculator utility functions to speculator.py
