+++ fms_extras/utils/cache/__init__.py
@@ -243,92 +243,3 @@ class KVCacheManager(metaclass=abc.ABCMeta):
 
 
 KVCache = Tuple[torch.Tensor, torch.Tensor]  # (key cache, value cache)
-
-
-def flatten_batch(inp: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-    """
-    Takes a speculator suffix tree: a bsize x n_candidates x candidate_len rectangular batch
-    of token indices, and flattens it while removing redundant tokens.
-
-    For example, given:
-
-    a b c
-    a b d
-    a e f
-
-    Tokens 'a b' in line 2 and token 'a' in line 3 are functionally equivalent to 'a b' in
-    line 1, so the flattened batch returns `a b c d e f`
-
-    Args:
-        inp: torch.Tensor
-            speculator suffix tree
-
-    Returns:
-        Tuple[torch.Tensor, torch.Tensor, torch.Tensor]
-            1) the flattened, pruned input
-            2) a tensor, sized as input, mapping each input token to its slot in output
-            3) a tensor, sized as output, mapping each output token to its slot in the flattened input
-    """
-    ind_out = torch.zeros_like(inp)
-    inp = inp.tolist()
-    out = []
-    ind_flat = []
-    batch_offset = 0
-    for b, candidate_set in enumerate(inp):
-        lineages = []
-        for k, candidate in enumerate(candidate_set):
-            for n in range(len(candidate)):
-                lineage = tuple(candidate[: n + 1])
-                if lineage in lineages:
-                    # Token is redundant
-                    ind_out[b, k, n] = lineages.index(lineage) + batch_offset
-                else:
-                    # Token is not redundant
-                    ind_out[b, k, n] = len(lineages) + batch_offset
-                    lineages.append(lineage)
-                    ind_flat.append(
-                        b * len(inp[0]) * len(inp[0][0]) + k * len(inp[0][0]) + n
-                    )
-        out.append(
-            torch.tensor(
-                [lineage[-1] for lineage in lineages],
-                device=ind_out.device,
-                dtype=torch.int32,
-            )
-        )
-        batch_offset += len(lineages)
-    return (
-        torch.cat(out),
-        ind_out,
-        torch.tensor(ind_flat, device=ind_out.device, dtype=torch.int32),
-    )
-
-
-def select_inflate_dim(
-    inp: torch.Tensor, inds: torch.Tensor, dim: int = 0
-) -> torch.Tensor:
-    """
-    Takes an input of size ([...], n, [...]), with n in slot corresponding to value of dim,
-    and tensor of indices of size (a, ..., z). Using those indices we over/under sample the
-    input on dimension n, to create output tensor with size ([...], (a, ..., z), [...]).
-
-    i.e. if dim=0, inp has size (6,3,2), and inds has size (8,4), then:
-    1) max(inds) < 6
-    2) output has size (8,4,3,2)
-
-    Args:
-        inp: torch.Tensor
-            tensor of inputs
-        inds: torch.Tensor
-            tensor of indices
-        dim: int
-            dimension to sample on
-
-    Returns:
-        torch.Tensor
-            output tensor with new size ([...], (a, ..., z), [...])
-    """
-    inds_shape = inds.size()
-    inp_shape = inp.size()
-    out = inp.index_select(dim, inds.view(-1))
-    return out.view(*inp_shape[:dim], *inds_shape, *inp_shape[dim + 1 :])
